Add ceph-ansible library and callback path
The ceph-ansible project has created a new library which
is held in the root folder. We need to ensure that the
library path is included.

Additionally a callback plugins directory inside plugins/callback has
been added, along with an action plugins directory inside
plugins/actions.

Finally, the sed command we use to replace the default roles directory
was causing issues in that it applied to any lines that were not
necessarily the roles_path. This PR adjusts the sed to ensure it is only
the roles_path line that is replaced.

Change-Id: I0d30d749ef2e00c1d199e53a837462619a9b2650